+++ ArtSpaceDos/Controllers/ArtDetailViewController.swift
