--- ArtSpaceDos/Utilities/ReuseIdentifiers.swift
@@ -1,14 +1,0 @@
-//
-//  ReuseIdentifiers.swift
-//  ArtSpaceDos
-//
-//  Created by Levi Davis on 2/18/20.
-//  Copyright © 2020 Jocelyn Boyd. All rights reserved.
-//
-
-import Foundation
-
-enum ReuseIdentifier: String {
-    case artCell
-    case savedArtCell
-}
